Add compiler flags.
* Turn on the long-disabled `-Ywarn-numeric-widen` after eliminating the
  warnings and
* add options to take advantage of Java 8 features (e.g., invoke dynamic
  & lambdas).

This requires a few other changes for compilation to work.

* Convert some types to the preferred Prism style,
* add `pPrism` to the `fp` package to construct a Prism from a partial
  function, and
* build docs from non-coverage-instrumented code.

+++ core/src/main/scala/quasar/physical/mongodb/util.scala
@@ -77,7 +77,7 @@ object util {
       */
     def testConnection(aclient: AMongoClient): Task[Unit] =
       MongoDbIO.serverVersion.run(aclient)
-        .timed(defaultTimeoutMillis)(Strategy.DefaultTimeoutScheduler)
+        .timed(defaultTimeoutMillis.toLong)(Strategy.DefaultTimeoutScheduler)
         .attempt flatMap {
           case -\/(tout: TimeoutException) =>
             val hosts = aclient.getSettings.getClusterSettings.getHosts.toString
